+++ packages/chookscord/src/definitions.ts
@@ -1,5 +1,6 @@
+/* eslint-disable object-curly-newline */
 import type * as types from '@chookscord/types';
-import type { Config, Event } from './types';
+import type { ChooksLifecycle, Config, Event, EventContext } from './types';
 import type { ClientEvents } from 'discord.js';
 
 export function defineConfig(
@@ -8,19 +9,23 @@ export function defineConfig(
   return config;
 }
 
-export function defineEvent<T extends keyof ClientEvents>(
-  event: Event<T>,
-): Event<T> {
+export function defineEvent<T extends keyof ClientEvents, Deps extends Record<string, unknown>>(
+  event: Omit<Event<T, Deps>, 'execute'> &
+  { execute: (ctx: EventContext, ...args: ClientEvents[T]) => unknown } &
+  ThisType<Readonly<Deps>>,
+): Event<T, Deps> {
   return event;
 }
 
-export function defineCommand(
-  command: types.ChooksSlashCommand,
-): types.ChooksSlashCommand {
+export function defineSlashCommand<T extends Record<string, unknown>>(
+  command: Omit<types.ChooksSlashCommand<T>, 'execute'> &
+  { execute: (ctx: types.ChooksCommandContext) => unknown } &
+  ThisType<Readonly<T>>,
+): types.ChooksSlashCommand<T> {
   return command;
 }
 
-export function defineSubCommand(
+export function defineSlashSubCommand(
   command: types.ChooksSubCommand,
 ): types.ChooksSubCommand {
   return command;
@@ -30,4 +35,30 @@ export function defineContextCommand(
   command: types.ChooksContextCommand,
 ): types.ChooksContextCommand {
   return command;
+}
+
+export function defineNonCommandOption(
+  option: types.ChooksNonCommandOption,
+): types.ChooksNonCommandOption {
+  return option;
+}
+
+export function defineSubCommandGroup(
+  group: types.ChooksGroupCommandOption,
+): types.ChooksGroupCommandOption {
+  return group;
+}
+
+export function defineSubCommand<T extends Record<string, unknown>>(
+  subCommand: Omit<types.ChooksSubCommandOption<T>, 'execute'> &
+  { execute: (ctx: types.ChooksCommandContext) => unknown } &
+  ThisType<Readonly<T>>,
+): types.ChooksSubCommandOption<T> {
+  return subCommand;
+}
+
+export function defineLifecycle(
+  lifecycle: ChooksLifecycle,
+): ChooksLifecycle {
+  return lifecycle;
 }