+++ packages/chookscord/src/index.ts
@@ -1,2 +1,3 @@
+export * from '@chookscord/types';
 export * from './types/public';
 export * from './definitions';