+++ packages/chookscord/src/scripts/dev/index.ts
@@ -2,12 +2,41 @@ process.env.NODE_ENV = 'development';
 import * as lib from '@chookscord/lib';
 import * as tools from '../../tools';
 import * as utils from '../../utils';
+import { ChooksTeardownList, reload } from './load-files';
+import type { ChooksContext } from '@chookscord/types';
+import type { Client } from 'discord.js';
 import type { ModuleName } from '../../types';
+import { basename } from 'path';
 import { createModuleLoader } from './load-modules';
 import { createWatchCompiler } from './compiler';
 import { loadConfig } from './load-config';
+import { unloadModule } from './unload';
 
 const logger = lib.createLogger('[cli] Chooks');
+const store: ChooksTeardownList = new Map();
+
+function reloadFile(client: Client, filePath: string) {
+  const fileName = basename(filePath);
+  const ctx: ChooksContext = {
+    client,
+    fetch: lib.fetch,
+    logger: lib.createLogger(`[file] ${fileName}`),
+  };
+  reload(ctx, store, filePath, logger);
+}
+
+function loadDir(client: Client, dirName: string) {
+  createWatchCompiler({
+    root: utils.appendPath.fromRoot(),
+    input: dirName,
+    output: `.chooks/${dirName}`,
+    compile(filePath) {
+      for (const cacheId of unloadModule(filePath)) {
+        reloadFile(client, cacheId);
+      }
+    },
+  });
+}
 
 function findFiles() {
   return tools.findProjectFiles(
@@ -46,14 +75,11 @@ export async function run(): Promise<void> {
 
   for (const moduleName of projectFiles) {
     if (isModule(moduleName)) {
+      logger.debug(`Module "${moduleName}" loaded.`);
       loadModule(moduleName);
     } else {
-      // eslint-disable-next-line no-new
-      createWatchCompiler({
-        root: utils.appendPath.fromRoot(),
-        input: moduleName,
-        output: `.chooks/${moduleName}`,
-      });
+      logger.debug(`Local dir "${moduleName}" loaded.`);
+      loadDir(client, moduleName);
     }
   }