+++ packages/chookscord/src/scripts/dev/listeners/interactions.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
 import * as lib from '@chookscord/lib';
 import * as tools from '../../../tools';
 import type { Client } from 'discord.js';
@@ -10,18 +11,19 @@ export function attachInteractionListener(
   store: lib.Store<CommandModule>,
   logger?: Consola,
 ): void {
-  client.on('interactionCreate', interaction => {
+  client.on('interactionCreate', async interaction => {
     const commandName = resolveInteractionKey(interaction);
     if (!commandName) return;
 
-    const command = store.get(commandName);
-    if (!command) {
+    const mod = store.get(commandName);
+    if (!mod) {
       logger?.warn(`Command "${commandName}" was executed, but no handlers were registered.`);
       return;
     }
 
     const ctx = tools.getCommandCtx(client, commandName, interaction);
-    const execute = command.execute.bind(command, ctx);
+    const deps = await mod.target.dependencies?.call(undefined) ?? {};
+    const execute = mod.target.execute!.bind(deps, ctx);
 
     tools.executeCommand(commandName, execute);
   });