+++ packages/chookscord/src/scripts/start/load-modules.ts
@@ -5,7 +5,9 @@ import type { ModuleContext, ModuleName } from '../../types';
 import { executeCommand, getCommandCtx } from '../../tools';
 import { Client } from 'discord.js';
 
-export type CommandHandler = Exclude<types.ChooksCommand['execute'], undefined>;
+export type CommandHandler = (
+  ctx: types.ChooksCommandContext | types.ChooksContextCommandContext
+) => unknown;
 
 const logger = lib.createLogger('[cli] Modules');
 
@@ -59,7 +61,7 @@ export function createModuleLoader(
       } return;
       case 'subcommands': {
         const store = getCommandStore();
-        const { getSubCommands } = await import('./modules/sub-commands');
+        const { getSubCommands } = await import('./modules/subcommands');
         for await (const [key, command] of getSubCommands(modulePath)) {
           store.set(key, command.execute);
         }
@@ -68,7 +70,7 @@ export function createModuleLoader(
         const store = getCommandStore();
         const { getContextCommands } = await import('./modules/context-commands');
         for await (const [key, command] of getContextCommands(modulePath)) {
-          store.set(key, command.execute);
+          store.set(key, command.execute as CommandHandler);
         }
       }
     }