+++ packages/chookscord/src/scripts/start/modules/context-commands.ts
@@ -5,11 +5,11 @@ import { basename } from 'path';
 
 const logger = lib.createLogger('[cli] ContextCommands');
 
-// @todo(Choooks22): Bind dependencies to 'this'
-function prepareCommand(
+async function prepareCommand(
   command: ChooksContextCommand,
-): [key: string, command: ChooksContextCommand] {
-  const execute = command.execute.bind(command);
+): Promise<[key: string, command: ChooksContextCommand]> {
+  const deps = await command.dependencies?.call(undefined) ?? {};
+  const execute = command.execute.bind(deps);
   return [command.name, { ...command, execute }];
 }