+++ packages/chookscord/src/tools/command.ts
@@ -1,20 +1,20 @@
+import type { ChooksCommandContext, ChooksContextCommandContext } from '@chookscord/types';
 import type { Client, Interaction } from 'discord.js';
 import { createLogger, fetch } from '@chookscord/lib';
-import type { ChooksContext } from '@chookscord/types';
 import type { Consola } from 'consola';
 import { createTimer } from '../utils';
 
-export function getCommandCtx(
+export function getCommandCtx<T extends ChooksContextCommandContext | ChooksCommandContext>(
   client: Client,
   commandName: string,
   interaction: Interaction,
-): ChooksContext {
+): T {
   return {
     fetch,
     client,
     interaction,
     logger: createLogger(`[commands] ${commandName}`),
-  };
+  } as T;
 }
 
 export async function executeCommand(