+++ packages/chookscord/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './config';
 export * from './events';
+export * from './lifecycle';
 export * from './modules';