+++ packages/chookscord/src/utils/debounce.ts
@@ -1,5 +1,6 @@
 /* eslint-disable @typescript-eslint/no-shadow */
-import type { FlattenPromise } from './types';
+import * as timers from 'timers/promises';
+import type { TimerOptions } from 'timers';
 
 export function debounce<T extends [...args: unknown[]]>(
   fn: (...args: T) => unknown,
@@ -24,31 +25,44 @@ export function debounce<T extends [...args: unknown[]]>(
   };
 }
 
+export type Debouced<T> =
+  { data: T; aborted: false } |
+  { data: null; aborted: true };
+
 export function debounceAsync<T extends [...args: unknown[]], R>(
   fn: (...args: T) => R,
-  ms: number
-): (...args: T) => FlattenPromise<R>;
+  delay: number
+): (...args: T) => Promise<Debouced<Awaited<R>>>;
 export function debounceAsync<T extends [...args: unknown[]], R>(
   fn: (...args: T) => R,
-  ms: number,
+  delay: number,
   ...args: T
-): () => FlattenPromise<R>;
+): () => Promise<Debouced<Awaited<R>>>;
 export function debounceAsync<T extends [...args: unknown[]], R>(
   fn: (...args: T) => R,
-  ms: number,
+  delay: number,
   ...args: T
-): (...args: T) => FlattenPromise<R> {
-  let timeout: NodeJS.Timeout;
-  return (...a) => {
-    if (timeout) {
-      clearTimeout(timeout);
+): (...args: T) => Promise<Debouced<R>> {
+  let controller: AbortController;
+
+  // eslint-disable-next-line complexity
+  return async (..._args) => {
+    if (controller) {
+      controller.abort();
     }
-    return new Promise(res => {
-      timeout = setTimeout(
-        (...args: T) => { res(fn(...args)) },
-        ms,
-        ...args.length ? args : a,
-      );
-    }) as FlattenPromise<R>;
+
+    controller = new AbortController();
+    const options: TimerOptions = { signal: controller.signal };
+
+    try {
+      const cb = await timers.setTimeout(delay, fn, options);
+      const data = await cb(...args.length ? args : _args);
+      return { data, aborted: false };
+    } catch (error) {
+      if (error instanceof Error && error.name === 'AbortError') {
+        return { data: null, aborted: true };
+      }
+      throw error;
+    }
   };
 }