+++ packages/chookscord/src/utils/index.ts
@@ -3,4 +3,3 @@ export * from './append-path';
 export * from './debounce';
 export * from './store-keys';
 export * from './timer';
-export * from './types';